refactor(logging): migrate capability validation to Arikawa and decouple feature flag evaluation from Discord intents

+++ pkg/control/feature_workspace_builder.go
@@ -299,8 +299,8 @@ func buildLogFeatureDetails(
 
 	out := &featureDetails{
 		RequiresChannel:         capability.RequiresChannel,
-		RequiredIntentsMask:     capability.RequiredIntentsMask,
-		RequiredPermissionsMask: capability.RequiredPermsMask,
+		RequiredIntentsMask:     int(capability.RequiredIntentsMask),
+		RequiredPermissionsMask: int64(capability.RequiredPermsMask),
 		ValidateChannelPerms:    capability.ValidateChannelPerms,
 		ExclusiveModeration:     capability.RequireExclusiveModeration,
 	}
